+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 <!-- SPDX-License-Identifier: Apache-2.0 -->
 
+## Unreleased
+
+## Added
+
+* Add `BitcrowdEcto.Changeset.validate_date_after/3` Validates a date field in the changeset is after the given reference date.
+* Add `BitcrowdEcto.Changeset.validate_hex_color/2` validates a changeset field has hexadecimal color format.
 
 ## Unreleased 
 
@@ -9,6 +15,7 @@
 * Add `BitcrowdEcto.Assertions.assert_changeset_valid/1`.
 
 * Add `BitcrowdEcto.Assertions.refute_changeset_valid/1`.
+
 
 ## [0.10.0] - 2022-04-28